+++ proxy/src/main/java/com/jhl/common/pojo/ConnectionLimit.java
@@ -1,0 +1,11 @@
+package com.jhl.common.pojo;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class ConnectionLimit {
+
+    private String accountNo;
+}
