+++ proxy/src/main/java/com/jhl/common/pojo/ProxyAccountWrapper.java
@@ -1,0 +1,19 @@
+package com.jhl.common.pojo;
+
+import com.ljh.common.model.ProxyAccount;
+import lombok.*;
+
+@EqualsAndHashCode(callSuper = true)
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class ProxyAccountWrapper extends ProxyAccount {
+    /**
+     * 版本号
+     */
+    private  Long version;
+
+
+
+}
