--- proxy/src/main/java/com/jhl/pojo/ConnectionLimit.java
@@ -1,11 +1,0 @@
-package com.jhl.pojo;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class ConnectionLimit {
-
-    private String accountNo;
-}
