--- proxy/src/main/java/com/jhl/pojo/ProxyAccountWrapper.java
@@ -1,19 +1,0 @@
-package com.jhl.pojo;
-
-import com.ljh.common.model.ProxyAccount;
-import lombok.*;
-
-@EqualsAndHashCode(callSuper = true)
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class ProxyAccountWrapper extends ProxyAccount {
-    /**
-     * 版本号
-     */
-    private  Long version;
-
-
-
-}
