+++ updated-log.md
@@ -1,4 +1,6 @@
 ## Next version
+#### Nagle算法禁用， TCP包_增加QOS高优先转发标识
+#### 复用端口
 ### V3.1.5
 #### 7.增加重置管理员管理命令
 #### 6.修改结算时间默认为30天
